backend/processing/color_flow.py
"""
Color Science Analysis — colour-science integration

Uses the colour-science library (the most comprehensive Python color science
library) for precise color analysis that the frontend HSL-based code cannot do:

  - Bradford chromatic adaptation (standard CIE illuminant adaptation)
  - Precise color temperature in Kelvin (not approximation from hue)
  - CIE ΔE2000 color differences
  - XYZ/xyY tristimulus conversions
  - Harmony detection
  - OKLCH-space analysis

This replaces the estimateColorTemperature function from imageAnalysis.ts
with physically accurate Kelvin computation.

Usage:
    result = analyze_image_colors("uploads/image.jpg")
    print(result["temperature_kelvin"])  # e.g., 5600.0 (daylight)
"""

import numpy as np
import colour
from PIL import Image
from sklearn.cluster import KMeans


def analyze_image_colors(image_path: str, n_colors: int = 5) -> dict:
    """
    Comprehensive color analysis using colour-science.

    Args:
        image_path: Path to the image file.
        n_colors: Number of dominant colors to extract.

    Returns:
        Dict with temperature_kelvin, dominant colors (hex + OKLCH),
        harmony assessment, and other metrics.
    """
    # Load and prepare image
    img = Image.open(image_path).convert("RGB")

    # Downsample for performance (max 500px on longest side)
    max_dim = 500
    if max(img.size) > max_dim:
        ratio = max_dim / max(img.size)
        new_size = (int(img.size[0] * ratio), int(img.size[1] * ratio))
        img = img.resize(new_size, Image.LANCZOS)

    pixels = np.array(img, dtype=np.float64) / 255.0  # normalize to 0–1
    flat_pixels = pixels.reshape(-1, 3)

    # -----------------------------------------------------------------------
    # 1. Color temperature estimation
    # -----------------------------------------------------------------------
    avg_rgb = flat_pixels.mean(axis=0)
    temperature_kelvin, temperature_delta = _estimate_cct(avg_rgb)

    # -----------------------------------------------------------------------
    # 2. Dominant color extraction (K-Means in sRGB → convert to OKLCH)
    # -----------------------------------------------------------------------
    kmeans = KMeans(n_clusters=n_colors, n_init=3, max_iter=100, random_state=42)
    kmeans.fit(flat_pixels)
    centroids = kmeans.cluster_centers_

    dominant_hex = []
    dominant_oklch = []

    for centroid in centroids:
        # sRGB → hex
        r, g, b = (np.clip(centroid * 255, 0, 255)).astype(int)
        hex_color = f"#{r:02x}{g:02x}{b:02x}"
        dominant_hex.append(hex_color)

        # sRGB → XYZ → OKLab → OKLCH
        oklch = _rgb_to_oklch(centroid)
        dominant_oklch.append({
            "l": round(float(oklch[0]), 4),
            "c": round(float(oklch[1]), 4),
            "h": round(float(oklch[2]), 2),
        })

    # -----------------------------------------------------------------------
    # 3. Image-wide metrics
    # -----------------------------------------------------------------------
    # Convert all pixels to OKLCH for perceptually accurate metrics
    avg_oklch = _rgb_to_oklch(avg_rgb)
    avg_lightness = float(avg_oklch[0])
    avg_chroma = float(avg_oklch[1])

    # Saturation level classification
    if avg_chroma < 0.03:
        saturation_level = "desaturated"
    elif avg_chroma < 0.08:
        saturation_level = "neutral"
    elif avg_chroma < 0.15:
        saturation_level = "moderate"
    else:
        saturation_level = "vivid"

    # Dynamic range (std dev of lightness)
    lightness_values = []
    sample_step = max(1, len(flat_pixels) // 1000)
    for i in range(0, len(flat_pixels), sample_step):
        oklch = _rgb_to_oklch(flat_pixels[i])
        lightness_values.append(oklch[0])
    dynamic_range = float(np.std(lightness_values)) * 100

    # -----------------------------------------------------------------------
    # 4. Color harmony detection
    # -----------------------------------------------------------------------
    hues = [c["h"] for c in dominant_oklch if c["c"] > 0.02]
    harmony = _detect_harmony(hues)

    return {
        "temperature_kelvin": round(temperature_kelvin, 1),
        "temperature_delta": round(temperature_delta, 2),
        "dominant_colors_hex": dominant_hex,
        "dominant_colors_oklch": dominant_oklch,
        "average_lightness": round(avg_lightness * 100, 2),
        "average_chroma": round(avg_chroma, 4),
        "color_harmony": harmony,
        "saturation_level": saturation_level,
        "dynamic_range": round(dynamic_range, 2),
    }


def _estimate_cct(srgb: np.ndarray) -> tuple[float, float]:
    """
    Estimate Correlated Color Temperature (CCT) using colour-science.

    This uses the proper CIE method:
      sRGB → XYZ → xy chromaticity → CCT (Robertson 1968 / Ohno 2014)

    Returns (kelvin, delta) where delta is the distance from the Planckian locus
    (positive = greenish tint, negative = magenta tint).
    """
    try:
        # Linearize sRGB
        linear = colour.cctf_decoding(srgb, function="sRGB")
        # Linear RGB → CIE XYZ
        xyz = colour.sRGB_to_XYZ(linear)
        # XYZ → xy chromaticity
        xy = colour.XYZ_to_xy(xyz)
        # xy → CCT using Ohno 2014 method
        cct, duv = colour.xy_to_CCT_Ohno2013(xy)
        return float(cct), float(duv)
    except Exception:
        # Fallback to approximate method
        return 5500.0, 0.0


def _rgb_to_oklch(srgb: np.ndarray) -> np.ndarray:
    """
    Convert sRGB (0–1) to OKLCH using colour-science.

    Pipeline: sRGB → Linear → XYZ → OKLab → OKLCH
    """
    try:
        linear = colour.cctf_decoding(np.clip(srgb, 0, 1), function="sRGB")
        xyz = colour.sRGB_to_XYZ(linear)

        # XYZ → OKLab (using colour-science's implementation)
        oklab = colour.XYZ_to_Oklab(xyz)

        # OKLab → OKLCH
        L = oklab[0]
        a = oklab[1]
        b_val = oklab[2]
        C = np.sqrt(a**2 + b_val**2)
        h = np.degrees(np.arctan2(b_val, a)) % 360

        return np.array([L, C, h])
    except Exception:
        return np.array([0.5, 0.0, 0.0])


def _detect_harmony(hues: list[float]) -> str:
    """
    Detect color harmony type from a list of OKLCH hues.

    Checks for: complementary, analogous, triadic, split-complementary, monochromatic.
    """
    if len(hues) < 2:
        return "monochromatic"

    # Calculate pairwise hue differences
    diffs = []
    for i in range(len(hues)):
        for j in range(i + 1, len(hues)):
            diff = abs(hues[i] - hues[j])
            if diff > 180:
                diff = 360 - diff
            diffs.append(diff)

    max_diff = max(diffs)
    min_diff = min(diffs)
    avg_diff = sum(diffs) / len(diffs)

    # Classification
    if max_diff < 30:
        return "monochromatic"
    elif max_diff < 60:
        return "analogous"
    elif 150 < max_diff < 210 and len(diffs) >= 1:
        # Check for split-complementary (one pair ~180°, others ~150° or ~30°)
        near_180 = sum(1 for d in diffs if 150 < d < 210)
        if near_180 >= 1 and min_diff < 60:
            return "split-complementary"
        return "complementary"
    elif 100 < avg_diff < 140:
        return "triadic"
    elif 70 < avg_diff < 100:
        return "tetradic"
    else:
        return "diverse"
