backend/requirements.txt
# ============================================================================
# VISIONFLOW Backend — Python Dependencies
# ============================================================================

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Image Processing
opencv-python-headless==4.10.0.84
Pillow==10.4.0
numpy>=1.26.4

# Color Science (Bradford chromatic adaptation, precise Kelvin, ΔE2000)
colour-science>=0.4.5

# Background Removal (U²-Net under the hood, zero config)
rembg>=2.0.62

# AI / Super-Resolution (ONNX for fast inference without full PyTorch)
onnxruntime>=1.19.0

# Machine Learning (K-Means clustering for dominant color extraction)
scikit-learn>=1.5.0

# Utilities
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.1
aiofiles==24.1.0

# Development & Testing
pytest==8.3.0
pytest-asyncio==0.24.0
httpx==0.27.0
black==24.8.0
ruff==0.6.0
